refactor: re-organise some modules
