+++ src/App.vue
@@ -137,8 +137,8 @@ import LangMenu from "@/components/menus/LangMenu.vue";
 import SaveMenu from "@/components/menus/SaveMenu.vue";
 import ThemeMenu from "@/components/menus/ThemeMenu.vue";
 import { Category, getPersistentModels } from "@/core/models";
-import { loadFromLegacyStorage } from "@/core/persistence/save-manager";
 import { FolderRepo } from "@/core/repositories";
+import { loadFromLegacyStorage } from "@/core/save/save-manager";
 import { t as $t } from "@/core/translation";
 import { VERSION } from "@/core/utils/types";
 import { usePreferencesStore } from "@/store/preferences";
