+++ src/store/preferences.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 import { ref, watch } from "vue";
 import { useTheme } from "vuetify";
-import { LocalStorageKey } from "@/core/persistence/save-manager";
+import { LocalStorageKey } from "@/core/save/save-manager";
 import { SupportedLanguages } from "@/core/translation/translation";
 import type { Theme } from "@/plugins/vuetify";
